+++ src/DaprExtension/Triggers/DaprBindingTriggerAttribute.cs
@@ -1,4 +1,7 @@
-﻿namespace Microsoft.Azure.WebJobs.Extensions.Dapr
+﻿// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace Microsoft.Azure.WebJobs.Extensions.Dapr
 {
     using System;
     using Microsoft.Azure.WebJobs.Description;
@@ -10,7 +13,7 @@
 #pragma warning disable CS0618 // Type or member is obsolete
     [Binding(TriggerHandlesReturnValue = true)]
 #pragma warning restore CS0618 // Type or member is obsolete
-    public class DaprInputBindingTriggerAttribute : Attribute
+    public class DaprBindingTriggerAttribute : Attribute
     {
         /// <summary>
         /// Gets or sets the name of the Dapr trigger.
@@ -18,6 +21,6 @@
         /// <remarks>
         /// If not specified, the name of the function is used as the trigger name.
         /// </remarks>
-        public string? TriggerName { get; set; }
+        public string? BindingName { get; set; }
     }
 }
